+++ CHANGELOG.md
@@ -5,6 +5,7 @@ REXpixel - Changelog
 
 * Fehler behoben
 * Sticky Position eingebaut
+* Sticky Position wieder ausgebaut. Tool ist jetzt verschiebbar und Sticky
 
 
 ### Version 0.2 - 07. Dezember 2013